@@ -3,24 +3,24 @@ description: >-
   Generator tool used with Rosetta annotations.
   This generates the helper and delegate classes
   for Flutter localizations.
-version: 0.2.2
+version: 0.3.0
 homepage: https://github.com/TeamWanari/rosetta
 
 environment:
-  sdk: ">=2.10.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
-  build: ">=1.5.0 <1.7.0"
-  source_gen: ">0.9.0 <0.11.1"
+  build: "^2.0.0"
+  source_gen: "^0.9.10+4"
   code_builder: ">3.3.0 <3.8.0"
-  dart_style: ">=1.0.0 <1.4.1"
-  analyzer: ">=0.40.0 <0.41.2"
-  recase: ">2.0.0 <3.1.0"
+  dart_style:  "^1.0.0"
+  analyzer: "^1.2.0"
+  recase: "^4.0.0-nullsafety.0"
   path: ">1.1.2 <1.8.1"
-  glob: ">1.1.0 <1.3.1"
+  glob: "^2.0.0"
   rosetta:
     path: ../rosetta/
 
 dev_dependencies:
-  build_runner: ">=1.8.0 <1.12.0"
+  build_runner: "^1.12.2"
   
