@@ -40,7 +40,7 @@ void checkTranslationKeyMap(List<Translation> translations) {
 
 void checkInterceptorFormat(MethodElement element) {
   var annotation = interceptorTypeChecker.firstAnnotationOfExact(element);
-  var isFiltered = annotation.getField("isFiltered").toBoolValue();
+  var isFiltered = annotation.getField("isFiltered")!.toBoolValue()!;
 
   if (element.returnType.getDisplayString(withNullability: false) != "String") {
     throw InvalidGenerationSourceError(
@@ -65,7 +65,7 @@ void checkInterceptorFormat(MethodElement element) {
           element: element);
     }
 
-    var regexpString = annotation.getField("filter").toStringValue();
+    var regexpString = annotation.getField("filter")!.toStringValue();
 
     if (regexpString == null) {
       throw InvalidGenerationSourceError(
